% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_to_dtm.R
\name{doc_to_dtm}
\alias{doc_to_dtm}
\title{Convert list of documents from stm package to sparse term-document matrix}
\usage{
doc_to_dtm(out, binary = TRUE)
}
\arguments{
\item{out}{A list of the output from stm \code{prepDocuments}.}

\item{binary}{A logical scalar. If TRUE (default) then only count one occurrence of a word in a
document.}
}
\description{
\code{doc_to_dtm} converts list of documents from stm package
\code{prepDocuments} function to a sparse term-document matrix.
}
\examples{

\dontrun{
library(stm)

processed <- textProcessor(
    input_data$text,
    data.frame(input_data),
    removestopwords=T, lowercase=T, stem=F
    )
out <- prepDocuments(
    processed$documents, processed$vocab, processed$meta
    )

ddm <- doc_to_dtm(out)
}

}
\seealso{
\code{\link[stm]{prepDocuments}},
\code{\link[stm]{textProcessor}},
\code{\link{scale_text_pcs}},
\code{\link{get_keywords}}, \code{\link{plot_keywords}},
\code{\link{score_documents_pcs}}
}
