% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keywords.R
\name{get_keywords}
\alias{get_keywords}
\title{List keywords from \code{scale_text}}
\usage{
get_keywords(
  scores,
  n_dimensions,
  n_words = 15,
  capture_output = FALSE,
  multiply_by_freq = TRUE
)
}
\arguments{
\item{scores}{List from output of \code{scale_text}.}

\item{n_dimensions}{Integer scalar or vector. How many/which dimensions to
print.}

\item{n_words}{An integer scalar. How many keywords for each dimension.}

\item{capture_output}{A logical scalar. Whether to return the output as list
of data frames rather than print to console.}

\item{multiply_by_freq}{A logical scalar. Whether to multiply the word scores by
the square root of the words' frequencies. Defaults to TRUE.}
}
\description{
\code{get_keywords} lists keywords for each scaled text dimension.
}
\examples{
\dontrun{
scores <- scale_text_pcs(
    meta=out$meta,
    dtm=dtm
    )

get_keywords(scores, n_dimensions=3, n_words=15)
}

}
\seealso{
\code{\link{plot_keywords}},
\code{\link{score_documents_pcs}}, \code{\link{doc_to_dtm}}
}
