% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labels.R
\name{get_labels}
\alias{get_labels}
\title{List labels from \code{scale_text_pcs}}
\usage{
get_labels(scores, n_dimensions, n_labels = 15, capture_output = FALSE)
}
\arguments{
\item{scores}{List from output of \code{scale_text}.}

\item{n_dimensions}{Integer scalar or vector. How many/which dimensions to
print.}

\item{n_labels}{Integer scale. How many labels to print.}

\item{capture_output}{A logical scalar. Whether to return the output as list
of data frames rather than print to console.}
}
\description{
\code{get_labels} lists labels for each scaled text dimension.
}
\examples{
\dontrun{
get_labels(scores, n_dimensions=3, n_labels=15)
}

}
\seealso{
\code{\link{scale_text_pcs}}
}
