% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{likes_dislikes_dtm}
\alias{likes_dislikes_dtm}
\title{ANES likes/dislikes document-term matrix}
\format{
## `likes_dislikes_dtm`
A sparseMatrix with 62,798 rows and 12,121 columns:
\describe{
  \item{rows}{Documents (a single open-ended response)}
  \item{columns}{Terms (whether a given word was used in a response)}
}
}
\usage{
likes_dislikes_dtm
}
\description{
A document-term matrix of open-ended survey responses to the question
"Is there anything in particular that you (like/dislike) about the
(Democratic/Republican) party? What is that?". As pre-processed for the paper
"Categorizing topics versus inferring attitudes: a theory and method for
analyzing open-ended survey responses."
}
\keyword{datasets}
