% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{likes_dislikes_meta}
\alias{likes_dislikes_meta}
\title{ANES likes/dislikes meta data}
\format{
## `likes_dislikes_meta`
A data frame with 62,798 rows and 5 columns:
\describe{
  \item{survey_year}{Survey year}
  \item{respondent_id}{Respondent ID}
  \item{question}{
  Indicates if a response explains like or dislike for the Democratic
  or Republican party.
  }
  \item{weight}{ANES survey weight, normalized by survey year}
  \item{holdout}{Whether the open-ended response was in the holdout set for the article.}
  ...
}
}
\usage{
likes_dislikes_meta
}
\description{
Meta data for the open-ended survey responses to the question
"Is there anything in particular that you (like/dislike) about the
(Democratic/Republican) party? What is that?". As formatted for the paper
"Categorizing topics versus inferring attitudes: a theory and method for
analyzing open-ended survey responses."
}
\keyword{datasets}
