% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_labels.R
\name{plot_labels}
\alias{plot_labels}
\title{List labels from \code{scale_text_pcs}}
\usage{
plot_labels(scores, dimension = 1)
}
\arguments{
\item{scores}{List from output of \code{scale_text}.}

\item{dimension}{Integer scalar or vector. Which dimension to
plot.}
}
\description{
\code{plot_labels} lists labels for each scaled text dimension.
}
\examples{
\dontrun{
plot_labels(scores, n_dimensions=1)
}

}
\seealso{
\code{\link{scale_text_pcs}}
}
