% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_documents_pcs.R
\name{score_documents_pcs}
\alias{score_documents_pcs}
\title{Score documents}
\usage{
score_documents_pcs(scores, max_dimensions = NULL)
}
\arguments{
\item{scores}{A list from output of \code{scale_text}. meta must be specified
in \code{scale_text_pcs} for this function to work.}

\item{max_dimensions}{An integer scalar. How many dimensions of scaled text to
score. Defaults to max_dimensions specified in \code{scale_text}}
}
\description{
\code{score_documents} uses word scores to assign document scores
}
\examples{

\dontrun{
scores <- scale_text_pcs(
    meta = out$meta,
    dtm = dtm
    )

document_scores <- score_documents_pcs(
    scores = scores, n_dimensions = 10
    )
}

}
\seealso{
\code{\link{scale_text_pcs}},
\code{\link{get_keywords}}, \code{\link{plot_keywords}},
\code{\link{doc_to_dtm}}
}
