% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_text_pcs.R
\name{scale_text_pcs}
\alias{scale_text_pcs}
\title{Scale text using implied words scaling}
\usage{
scale_text_pcs(
  dtm,
  meta = NULL,
  dtm_vocab = NULL,
  cutoff_exponent = 2,
  verbose = TRUE,
  holdout = NULL,
  max_dimensions = 4,
  weights = NULL,
  project_rare_words = TRUE
)
}
\arguments{
\item{dtm}{A sparseMatrix. Rows are documents and columns are vocabulary.}

\item{meta}{data.frame. Must line up with \code{dtm} etc. This is included only to
keep track of any accompanying variables. It is unaltered by the function.}

\item{dtm_vocab}{A character vector. Provide vocabulary for columns of \code{dtm} if
missing in column names.}

\item{cutoff_exponent}{An integer scalar. Used to determine the number of
'common words'. Defaults to 2 (as in paper).}

\item{verbose}{A logical scalar. Print progress of the function.}

\item{holdout}{A logical vector, numeric vector, or character scalar. A
logical or numeric vector indicates which rows to exclude from training. A
character scalar provides the name of a column in \code{meta}.}

\item{max_dimensions}{An integer scalar. The number of principal components
to calculate.}

\item{weights}{A numeric vector or character scalar. Responses weights.
If a numeric vector is provided, it must be equal to the number of rows in
\code{dtm}. A character scalar provides the name of a column in \code{meta}.}

\item{project_rare_words}{A logical scalar. Whether to project rare words not (fully)
used in the scaling process.}
}
\description{
\code{scale_text_pcs} runs implied words scaling
}
